move the diff function into a separate module
